Refactor Python code
Combine import and search into one CLI
for simpler maintenance and usability.
