+++ python/main.py
@@ -1,40 +1,27 @@
 #!/usr/bin/python3
-
 import argparse
 import json
 import os
-import re
-import sys
-import requests
-import json
-import logging
 from pathlib import Path
-
+import re
+import requests
+from sentence_transformers import SentenceTransformer
+import sys
 import yaml
+import logging
+
 
 import config
-import utils
 import mappers
+import utils
 from validator import validate_with_custom_logic
-
-
-def extract_yaml_frontmatter(filepath):
-    with open(filepath, "r", encoding="utf-8") as file:
-        content = file.read()
-    pattern = r"^---\n(.*?)\n---\n"
-    match = re.search(pattern, content, re.DOTALL | re.MULTILINE)
-    if match:
-        payload = match.group(1)  # Return the captured YAML content
-        payload = yaml.safe_load(payload)
-        return payload
-    return None
 
 
 def clean_up_old_versions(json_dataset):
     for filename in os.listdir(config.datasets_dir):
         if filename.endswith(".md"):
             filepath = os.path.join(config.datasets_dir, filename)
-            frontmatter = extract_yaml_frontmatter(filepath)
+            frontmatter = utils.extract_yaml_frontmatter(filepath)
             json_dataset_id = json_dataset.get("dataset_id")
             markdown_dataset_id = frontmatter.get("dataset_id")
             if frontmatter and markdown_dataset_id == json_dataset_id:
@@ -57,14 +44,14 @@ def fetch_schema(schema_url, schema_path):
 
 def validate_json_with_schema(dataset_from_json, schema_url):
     if schema_url == config.schema_url_v3:
-        schema_path = f"{config.schema_path}/rdl-03.json"
+        schema_path = f"{config.python_path}/rdl-03.json"
         is_cached = os.path.isfile(schema_path)
         if not is_cached:
             # TODO: delete & replace hardcoded_schema_url with schema_url when v0.3 finalized
             hardcoded_schema_url = "https://raw.githubusercontent.com/GFDRR/CCDR-tools/refs/heads/main/_static/rdls_schema_v0.3.json"
             fetch_schema(hardcoded_schema_url, schema_path)
     else:
-        schema_path = f"{config.schema_path}/rdl-02.json"
+        schema_path = f"{config.python_path}/rdl-02.json"
         is_cached = os.path.isfile(schema_path)
         if not is_cached:
             fetch_schema(schema_url, schema_path)
@@ -117,7 +104,7 @@ def write_datasets_to_markdown(json_filepaths):
         with open(json_filepath, encoding=encoding) as input_file:
             datasets_json = json.load(input_file)
             for dataset in datasets_json["datasets"]:
-                first_link, *_ = dataset.get("links")
+                first_link, *_ = dataset.get("links", [{"href": config.schema_url_v2}])
                 schema_url = first_link.get("href")
                 result = write_dataset_to_markdown(dataset, schema_url)
                 if result != 0:
@@ -125,26 +112,7 @@ def write_datasets_to_markdown(json_filepaths):
     return exit_code
 
 
-if __name__ == "__main__":
-    # Parse args
-    parser = argparse.ArgumentParser(
-        description="Convert RDL JSON datasets into JKAN frontmatter"
-    )
-    parser.add_argument(
-        "--input_folder",
-        help="Path to the folder containing RDL datasets in JSON format",
-        default=".",
-        action="store",
-    )
-    parser.add_argument(
-        "--type",
-        choices=["ci", "batch"],
-        help="Whether the datasets are being generated from new/modified files from the last commit or from a batch directory",
-        default="batch",
-        action="store",
-    )
-    args = parser.parse_args()
-    # Create output paths if they don't already exist
+def check_for_required_paths():
     if not Path(config.root_dir).is_dir():
         os.makedirs(config.root_dir)
     if not Path(config.datasets_dir).is_dir():
@@ -152,25 +120,75 @@ if __name__ == "__main__":
     if not Path(config.json_dir).is_dir():
         os.makedirs(config.json_dir)
 
-    exit_code = 0
-    if args.type == "ci":
-        result = write_datasets_to_markdown(utils.get_recently_changed_json_files())
-        if result != 0:
-            exit_code = result
-    elif args.type == "batch":
-        input_path = Path(args.input_folder)
-        result = write_datasets_to_markdown(
-            input_path.glob(f"{config.json_dir}/*.json")
-        )
-        if result != 0:
-            exit_code = result
-    else:
-        raise ValueError(f"Unknown type {args.type}")
 
+def get_datasets_metadata():
+    metadata_entries = []
+    for filename in os.listdir(config.datasets_dir):
+        if filename.endswith(".md"):
+            filepath = os.path.join(config.datasets_dir, filename)
+            with open(filepath, "r", encoding="utf-8") as file:
+                content = file.read()
+                pattern = r"^---\n(.*?)\n---\n"
+                match = re.search(pattern, content, re.DOTALL | re.MULTILINE)
+                if match:
+                    payload = match.group(1)  # Return captured YAML
+                    payload = yaml.safe_load(payload)
+                    json_payload = json.dumps(payload)
+                    metadata_entries.append(json_payload)
+    return metadata_entries
+
+
+def embed_datasets_metadata(datasets_metadata):
+    results = []
+    for dataset_metadata in datasets_metadata:
+        tensor = model.encode(
+            dataset_metadata, convert_to_tensor=True, normalize_embeddings=True
+        )
+        vector_embedding = {
+            "vector": tensor.cpu().detach().numpy().tolist(),
+            "metadata": json.loads(dataset_metadata),
+        }
+        results.append(vector_embedding)
+
+    return results
+
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser(
+        description="Utility for importing RDL metadata into JKAN"
+    )
+    parser.add_argument(
+        "--ci",
+        help="Tells the command to only run on commits since the remote target branch defined in config.py",
+        action="store_true",
+    )
+    parser.add_argument(
+        "-v",
+        "--vectors",
+        help="Tells the command to generate vector embeddings of metadata",
+        action="store_true",
+    )
+    args = parser.parse_args()
+    check_for_required_paths()
+    files_to_process = (
+        utils.get_recently_changed_files(f"_datasets/json/*.json")
+        if args.ci
+        else Path(".").glob(f"{config.json_dir}/*.json")
+    )
+    exit_code = write_datasets_to_markdown(files_to_process)
+    if args.vectors is True and files_to_process:
+        model = SentenceTransformer("all-MiniLM-L6-v2")
+        datasets_metadata = get_datasets_metadata()
+        vector_embeddings = embed_datasets_metadata(datasets_metadata)
+
+        exit_code = exit_code | utils.save_to_json(
+            vector_embeddings, config.vectors_path
+        )
+        print(f"Vectors saved to {config.vectors_path}.")
     if exit_code == 0:
         print(
             "\nAll done! Please enjoy your datasets :)\n",
-            "More info is available at `import/README.md`\n",
+            "More info is available at `python/README.md`\n",
             sep=os.linesep,
         )
     else:
