+++ python/requirements.txt
@@ -3,3 +3,4 @@ GitPython==3.1.45
 jsonschema==4.25.1
 requests==2.32.5
 chardet==5.2.0
+sentence-transformers==5.1.2
