--- search/.gitignore
@@ -1,7 +1,0 @@
-# python
-bin
-lib
-lib64
-pyvenv.cfg
-.venv
-__pycache__
