--- search/config.py
@@ -1,3 +1,0 @@
-root_dir = ".."
-# input directory
-datasets_dir = f"{root_dir}/_datasets"
