--- search/requirements.txt
@@ -1,2 +1,0 @@
-sentence-transformers
-PyYAML==6.0.0
